requirements.txt
Flask==2.0.1
python-dotenv==0.19.0
google-generativeai==0.3.2
flask-cors==3.0.10
gunicorn==20.1.0
werkzeug==2.0.3 